+++ datafit/src/index.ts
@@ -5,4 +5,4 @@
  * Exports the public-facing API for `datafit`
  */
 export * from './types';
-export * from './CurveFit';
+export * from './lib';
